Add demo of job scheduling using quartz
The GreetingController is extended, so it now allows a POST which schedules a job to
run at a set frequency. The job just logs the Greeting to the console

Quartz is a framework which allows Jobs to be scheduled to run asynchronously, at a fixed
frequency or point in time. This is useful if youo want to allow the API user to
schedule a task to be performed at a set point in the future

The job schedule is persisted in a postgres database, so they will live beyond the lifespan
of an application run. I.e. the application can be restarted, and the running jobs will
restart.

Jira: n/a