Add Spring JPA example
Adds code to store and retrieve entities from a database using Spring JPA.

A BookRepository is added. Books and their authors are stored and retrieved using
the repository. This provides an example of usage of a join table, OneToMany and
OneToOne relationships in JPA so for creation and retrieval of entities.