Add endpoint to stream data read from database
Add a streaming http endpoint returning Book objects. Books are streamed from the database using
the JOOQ framework. The books read from the database are then streamed to an OutputStream.
Where there are a large number of documents needing to be returned, or processing each
document takes time, this allows the service to return documents to the client as then
are available, rather than having to wait for them all to be returned at the end

the endpoint can be tested with curl:

curl --no-buffer localhost:8080/books

where books can be seen streamed individually to the console.